pyproject.toml
[project]
name = "exam-helper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-openai",
    "langgraph",
    "python-dotenv",
    "pydantic",
    "sqlalchemy>=2.0.48",
    "pydantic-settings>=2.13.1",
    "fastapi",
    "uvicorn",
    "alembic>=1.18.4",
    "fastapi-users[sqlalchemy]>=15.0.5",
    "aiosqlite>=0.22.1",
]


[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project.scripts]
init-db = "scripts.init_db:init_db"
dev = "scripts.dev:run"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
norecursedirs = ["temp", ".*", "build", "dist", "node_modules", "venv", "tests/temp"]
